rearrange project and add pagination to playlists api request
